Изменил ID на BIGINT

+++ app/db/sqlalchemy/models/food.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, ForeignKey
-from sqlalchemy.dialects.postgresql import VARCHAR, INTEGER
+from sqlalchemy.dialects.postgresql import VARCHAR, BIGINT
 from sqlalchemy import Table
 from app.db.sqlalchemy.base import metadata_obj, mapper_registry
 from app.core.models.food import Food
@@ -8,9 +8,9 @@ from app.core.models.food import Food
 food = Table(
     "food",
     metadata_obj,
-    Column("id", INTEGER, primary_key=True),
+    Column("id", BIGINT, primary_key=True),
     Column("name", VARCHAR),
-    Column("category_id", INTEGER, ForeignKey("food_category.id")),
+    Column("category_id", BIGINT, ForeignKey("food_category.id")),
 )
 
 mapper_registry.map_imperatively(
